src/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.0
cffi==1.17.1
cryptography==43.0.3
fastapi==0.115.4
greenlet==3.1.1
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2

